Reorganize JournalTool classes
